Refactor to feature first
