+++ package.json
@@ -22,6 +22,7 @@
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
     "form-data": "^4.0.0",
+    "parse-torrent-name": "^0.5.4",
     "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
